# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReportResolution": {
        "additionalProperties": false,
        "properties": {
          "resolution": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Resolution",
            "type": "string"
          },
          "status": {
            "pattern": "^(triaged|investigating|resolved|dismissed|duplicate)$",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "resolution"
        ],
        "title": "ReportResolution",
        "type": "object"
      },
...
      "RoleAssignmentRequest": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          },
          "role": {
            "pattern": "^(contributor|reviewer|maintainer|administrator)$",
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "reason"
        ],
        "title": "RoleAssignmentRequest",
        "type": "object"
      },
...
      "UserModerationRequest": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          },
          "status": {
            "pattern": "^(active|suspended)$",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "reason"
        ],
        "title": "UserModerationRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Versioned, local-first API over deterministic generated knowledge contracts, strictly bounded graph intelligence, and proposal-only community collaboration. Community routes never mutate canonical knowledge; Git and maintainer review remain the publication boundary. Lab routes accept only predefined safe local-fixture actions; arbitrary commands and target scanning are not supported.",
    "title": "Universal Hacking Tools Knowledge API",
    "version": "10.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/community/admin/reports": {
      "get": {
        "operationId": "admin_reports_api_v1_community_admin_reports_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "uht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read private administrator report queue",
        "tags": [
          "community",
          "collaboration"
        ]
      }
    },
    "/api/v1/community/admin/reports/{report_id}/resolve": {
      "post": {
        "operationId": "admin_resolve_report_api_v1_community_admin_reports__report_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "uht_csrf",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Csrf"
            }
          },
          {
            "in": "cookie",
            "name": "uht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportResolution"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve a private report with audit logging",
        "tags": [
          "community",
          "collaboration"
        ]
      }
    },
    "/api/v1/community/admin/users/{user_id}/moderation": {
      "post": {
        "operationId": "admin_moderate_api_v1_community_admin_users__user_id__moderation_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "uht_csrf",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Csrf"
            }
          },
          {
            "in": "cookie",
            "name": "uht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserModerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Suspend or reactivate a user with audit logging",
        "tags": [
          "community",
          "collaboration"
        ]
      }
    },
    "/api/v1/community/admin/users/{user_id}/role": {
      "post": {
        "operationId": "admin_role_api_v1_community_admin_users__user_id__role_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "uht_csrf",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Csrf"
            }
          },
          {
            "in": "cookie",
            "name": "uht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleAssignmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assign a role with administrator-only audit logging",
        "tags": [
          "community",
          "collaboration"
        ]
      }
    },
    "/api/v1/community/contributions": {
      "get": {
        "operationId": "published_contributions_api_v1_community_contributions_get",
        "parameters": [
          {
            "in": "query",
            "name": "contribution_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Contribution Type"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 10000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List published community contributions only",
        "tags": [
          "community",
          "collaboration"
        ]
      }
    },
    "/api/v1/community/contributions/{contribution_id}": {
      "get": {
        "operationId": "published_contribution_api_v1_community_contributions__contribution_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "contribution_id",
            "required": true,
            "schema": {
              "title": "Contribution Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read one published community contribution",
        "tags": [
          "community",
          "collaboration"
        ]
      }
    },
    "/api/v1/community/contributors": {
      "get": {
        "operationId": "contributors_api_v1_community_contributors_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 40,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 10000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List public contributor profiles",
        "tags": [
          "community",
          "collaboration"
        ]
      }
    },
...
  }
}